rose-blanche-api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6
sentence-transformers>=2.2.0
numpy>=1.26.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
aiofiles>=24.1.0
tqdm>=4.66.0
celery>=5.4.0
flower>=2.0.0
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.21.0
PyPDF2>=3.0.0
